\name{ClonalFrameML-package}
\alias{ClonalFrameML-package}
\alias{ClonalFrameML}
\docType{package}
\title{
\packageTitle{ClonalFrameML}
}
\description{
\packageDescription{ClonalFrameML}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ClonalFrameML}
\packageIndices{ClonalFrameML}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{ClonalFrameML}

Maintainer: \packageMaintainer{ClonalFrameML}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
